requirements.txt
Flask==3.0.3
Jinja2==3.1.4
Werkzeug==3.0.3
itsdangerous==2.2.0
click==8.1.7
fpdf2==2.7.6

matplotlib==3.9.2
numpy==2.1.1

pandas==2.2.2
gunicorn==23.0.0
